+++ src/actions/hosts.ts
@@ -11,13 +11,20 @@ import { clearDNSCache } from './cache'
 
 export const saveHosts = async (context: vscode.ExtensionContext) => {
   const defaultHosts = await getDefaultHosts(context)
-const selectedConfig = getSelectedConfig(context)
-const fileString = selectedConfig ? `${defaultHosts}\n
+  const configSelected = getConfigSelected()
+  const selectedConfig = getSelectedConfig(configSelected)
+  const fileString = selectedConfig ? `${defaultHosts}\n
 # from switch-hosts
 ${selectedConfig}
 ` : defaultHosts
   
-  await writeElevated(path.sysHostsPathString, fileString).catch((err: Error) =>{
+  await writeElevated(path.sysHostsPathString, fileString).then(()=> {
+    vscode.window.showInformationMessage(`save hosts: ${configSelected.join(',')} sucess`, 'view').then(value => {
+      if (value === 'view') {
+        openSysHostsFile()
+      }
+    })
+  }, (err: Error) =>{
     console.log(err)
     vscode.window.showErrorMessage('write system hosts failed')
     return Promise.reject(err)
@@ -31,7 +38,18 @@ export const selectedConfig = async (context: vscode.ExtensionContext) => {
   const quickPickItem:vscode.QuickPickItem[] = Object.keys(configHosts)
     .map(host => ({label:host, picked: configSelected.includes(host)}))
 
-  const selected = await vscode.window.showQuickPick(quickPickItem, {canPickMany: true}) || []
+  const selected = await vscode.window.showQuickPick(quickPickItem,
+    {placeHolder: 'Select the configs to switch', canPickMany: true })
+  console.log(selected)
+  if (!selected) return
   await setConfigSelected(selected.map(pickItem=>pickItem.label))
   await saveHosts(context)
+}
+
+export const openSysHostsFile = () => {
+  vscode.workspace.openTextDocument(path.sysHostsPathString).then(document => {
+    vscode.window.showTextDocument(document);
+  }, error => {
+      return vscode.window.showErrorMessage("fastHosts: " + error.message);
+  });
 }