+++ src/extension.ts
@@ -13,7 +13,6 @@ export function activate(context: vscode.ExtensionContext) {
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
 	console.log('Congratulations, your extension "switch-hosts" is now active!');
-
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
@@ -38,7 +37,13 @@ export function activate(context: vscode.ExtensionContext) {
 		})
 	});
 
-	context.subscriptions.push(disposableSaveConfig, disposableSelectedConfig, disposableUpdateDefaultHosts);
+	let disposableOpenHostsFile = vscode.commands.registerCommand('switch-hosts.openHostsFile', async () => {
+		updateDefaultHosts(context).catch((err: Error)=> {
+			vscode.window.showErrorMessage(err.message)
+		})
+	});
+
+	context.subscriptions.push(disposableSaveConfig, disposableSelectedConfig, disposableUpdateDefaultHosts, disposableOpenHostsFile);
 }
 
 // this method is called when your extension is deactivated