+++ CHANGELOG.md
@@ -2,8 +2,27 @@
 
 All notable changes to the "switch-hosts" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [2.2.0] - 2020-03-25
 
-- Initial release
+* Support VS code Insiders version
+* add open system hosts file:  
+```Ctrl + Shift + p ```> ``` switch-hosts: openHostsFile ```
+* add saved hosts notification
+
+## 2.1.0
+
+* hosts update default hosts
+* add README features gif
+
+## 2.0.0
+
+Complete basic functions:
+
+* hosts config
+* hosts save
+* hosts selected
+
+## 1.0.0
+
+Initial release of switch-hosts